Add basic validation to the CSP details
There are several Go libraries that let you use tags on a struct to
specify ways in which the struct can be validated.

However, of the 2 that I've tried neither has validated things as I
would expect.

- https://github.com/asaskevich/govalidator
- https://github.com/go-validator/validator

The logic in this commit is not particularly nice, but for a first
attempt it does at least work.